v3

+++ services/warehouse-service/src/main/java/com/marko/logistics/warehouse/presentation/controller/WarehouseController.java
@@ -5,9 +5,14 @@ import com.marko.logistics.warehouse.application.dto.TotalWarehousesResponse;
 import com.marko.logistics.warehouse.application.dto.UpdateWarehouseRequest;
 import com.marko.logistics.warehouse.application.dto.WarehouseResponse;
 import com.marko.logistics.warehouse.application.service.WarehouseService;
+import com.marko.logistics.warehouse.infrastructure.security.RequestContext;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.UUID;
@@ -19,6 +24,7 @@ import java.util.UUID;
 public class WarehouseController {
 
     private final WarehouseService warehouseService;
+    private final RequestContext requestContext;
 
     @PostMapping
     public WarehouseResponse create(@RequestBody CreateWarehouseRequest request) {
@@ -29,7 +35,16 @@ public class WarehouseController {
     }
 
     @PutMapping("/{id}")
-    public WarehouseResponse update(@PathVariable UUID id, @RequestBody UpdateWarehouseRequest request) {
+    public WarehouseResponse update(
+            @PathVariable UUID id,
+            @RequestBody UpdateWarehouseRequest request,
+            HttpServletRequest httpRequest) {
+        String role = requestContext.getUserRole(httpRequest);
+
+        if(!"MANAGER".equalsIgnoreCase(role)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Managers only");
+        }
+
         log.info("Updating warehouse {} with data: {}", id, request.name());
         var response = warehouseService.updateWarehouse(id, request);
         log.info("Updated warehouse {} successfully", id);
@@ -59,7 +74,15 @@ public class WarehouseController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable UUID id) {
+    public void deleteById(
+            @PathVariable UUID id,
+            HttpServletRequest request) {
+        String role = requestContext.getUserRole(request);
+
+        if(!"MANAGER".equalsIgnoreCase(role)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Managers only");
+        }
+
         log.info("Deleting warehouse with ID: {}", id);
         warehouseService.deleteWarehouseById(id);
         log.info("Deleted warehouse {}", id);
